refactor: :recycle: organized context code
