+++ src/context/JSONContext.tsx
@@ -1,7 +1,8 @@
-import React, { createContext, useState, useReducer, useEffect } from "react";
+import React, { createContext } from "react";
 import { IFireboltJSON, IStep } from "@/types/fireboltJSON";
-import blankJSON from "./blankJSONBoilerplate";
-import reducer, { JSONAction } from "./reducer";
+
+import { JSONAction } from "./reducer";
+import useJSONContext from "./hook";
 
 interface IJSONProviderValues {
   currentJSON: IFireboltJSON;
@@ -13,35 +14,15 @@ interface IJSONProviderValues {
 export const JSONContext = createContext({} as IJSONProviderValues);
 
 export function JSONProvider({ children }) {
-  const [currentJSON, dispatch] = useReducer(reducer, blankJSON);
-  const [visibleStepState, setVisibleStepState] = useState<IStep>(
-    currentJSON?.steps[0]
-  );
-  const [visibleStepSlug, setVisibleStepSlug] = useState<string>(
-    currentJSON?.steps[0]?.step?.slug
-  );
-
-  useEffect(() => {
-    const stepData = currentJSON.steps.find(
-      (step) => step?.step?.slug === visibleStepSlug
-    );
-    if (stepData) setVisibleStepState(stepData);
-  }, [visibleStepSlug, currentJSON]);
-
-  useEffect(() => {
-    console.log(currentJSON);
-  }, [currentJSON]);
-
-  function setVisibleStep(step: IStep) {
-    setVisibleStepSlug(step?.step?.slug);
-  }
+  const { currentJSON, dispatch, visibleStep, setVisibleStep } =
+    useJSONContext();
 
   return (
     <JSONContext.Provider
       value={{
         currentJSON,
         dispatch,
-        visibleStep: visibleStepState,
+        visibleStep,
         setVisibleStep,
       }}
     >
